requirements.txt
pandas==1.5.3
pymongo==4.3.3
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
selenium==4.8.3
UliPlot==0.2.2
